refactor(print): rename files for clarity
- writer_runtime_os.cpp -> writer_runtime.cpp
  The file contains no OS-specific code; it implements CodeReader,
  buffer helpers, and float formatting utilities that are fully
  platform-agnostic.

- printf/printf_frontend_lowering_{base,field}.hpp
  -> printf/printf_frontend_binding_{base,field}.hpp
  The brace frontend calls the same concept 'binding' (binding a
  parsed field to a FormatType + pack combination). Using 'binding'
  consistently across both frontends reduces cross-reading friction.
